0.8.19中uint <-> int强制转换

+++ hardhat/contracts/factory/libraries/Tick.sol
@@ -43,7 +43,7 @@ library Tick {
     /// @return The max liquidity per tick
     function tickSpacingToMaxLiquidityPerTick(int24 tickSpacing) internal pure returns (uint128) {
         int24 minTick = (TickMath.MIN_TICK / tickSpacing) * tickSpacing;
-        int24 maxTick = (TickMath.MAX_TICK / tickSpacing) * tickSpacing;
+        int24 maxTick = (int24(TickMath.MAX_TICK) / tickSpacing) * tickSpacing;
         uint24 numTicks = uint24((maxTick - minTick) / tickSpacing) + 1;
         return type(uint128).max / numTicks;
     }
